requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2
googlemaps
geopy
python-decouple
requests